pyproject.toml
[project]
name = "viz-api"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "einops>=0.8.0",
    "pillow>=11.0.0",
    "litserve>=0.2.13",
    "python-multipart>=0.0.20",
    "hf-transfer>=0.1.8",
    "hf-xet>=1.0.3",
    "transformers>=4.47.1",
    "torch>=2.6.0",
]

[project.optional-dependencies]
cpu = [
    "torch>=2.7.0",
]

cu118 = [
    "torch>=2.6.0",
]

cu122 = [
    "torch>=2.6.0",
]

cu126 = [
    "torch>=2.7.0",
]

[dependency-groups]
dev = [
    "requests>=2.32.3",
    "black>=24.10.0",
    "ipython>=8.31.0",
    "isort>=5.13.2",
]

[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu118" },
    { extra = "cu122" },
    { extra = "cu126" },
  ],
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu",   extra = "cpu" },
    { index = "pytorch-cu118", extra = "cu118" },
    { index = "pytorch-cu124", extra = "cu122" },
    { index = "pytorch-cu126", extra = "cu126" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
